Moving dump method to NativeMethods
