+++ minuet/Paragon/Program.cs
@@ -29,65 +29,13 @@ using System.Runtime.InteropServices;
 using System.Collections.Generic;
 using System.IO.Packaging;
 using System.Text;
+using Paragon.Runtime.Win32;
+using Paragon.Runtime.Desktop;
 
 namespace Paragon
 {
     static class Program
     {
-        public class Dump
-        {        
-            internal enum MINIDUMP_TYPE
-            {
-                MiniDumpNormal = 0x00000000,
-                MiniDumpWithDataSegs = 0x00000001,
-                MiniDumpWithFullMemory = 0x00000002,
-                MiniDumpWithHandleData = 0x00000004,
-                MiniDumpFilterMemory = 0x00000008,
-                MiniDumpScanMemory = 0x00000010,
-                MiniDumpWithUnloadedModules = 0x00000020,
-                MiniDumpWithIndirectlyReferencedMemory = 0x00000040,
-                MiniDumpFilterModulePaths = 0x00000080,
-                MiniDumpWithProcessThreadData = 0x00000100,
-                MiniDumpWithPrivateReadWriteMemory = 0x00000200,
-                MiniDumpWithoutOptionalData = 0x00000400,
-                MiniDumpWithFullMemoryInfo = 0x00000800,
-                MiniDumpWithThreadInfo = 0x00001000,
-                MiniDumpWithCodeSegs = 0x00002000
-            }
-            [DllImport("dbghelp.dll" )]
-            static extern bool MiniDumpWriteDump (
-                IntPtr hProcess,
-                Int32 ProcessId,
-                IntPtr hFile,
-                MINIDUMP_TYPE DumpType,
-                IntPtr ExceptionParam,
-                IntPtr UserStreamParam,
-                IntPtr CallackParam );
-
-            public static void MiniDumpToFile ( String fileToDump, string processToDump )
-            {
-                FileStream fsToDump = null;
-                if ( File.Exists( fileToDump ) )
-                    fsToDump = File.Open( fileToDump, FileMode.Append );
-                else
-                    fsToDump = File.Create( fileToDump );
-
-                try
-                {
-                    Process process = Process.GetProcessesByName(processToDump).First();
-
-                    MiniDumpWriteDump(process.Handle, process.Id,
-                        fsToDump.SafeFileHandle.DangerousGetHandle(), MINIDUMP_TYPE.MiniDumpNormal,
-                        IntPtr.Zero, IntPtr.Zero, IntPtr.Zero);
-                    fsToDump.Close();
-                }
-                catch (Exception ex)
-                {
-                    throw (ex);
-                }
-            }
-        };
-        
         private static ApplicationManager _appManager;
 
         [STAThread]
@@ -279,18 +227,19 @@ namespace Paragon
                 String fullPath = Path.Combine(dstDiretory, fileName);
 
                 //Dump apps.
-                //This is not ok. Somehow, current available code to dump running apss is not ok.
-                //WebApplication runningApp = (WebApplication)_appManager.AllApplicaions.FirstOrDefault();
-                //Paragon.Runtime.Win32.MemoryDump.CreateMemoryDump((AppInfo)runningApp.GetRunningApps().FirstOrDefault(), Runtime.Win32.ProcessDumpType.MiniDumpWithFullMemory);
+                WebApplication runningApp = (WebApplication)_appManager.AllApplicaions.FirstOrDefault();
+                AppInfo appInfo = (AppInfo)runningApp.GetRunningApps().FirstOrDefault();
 
                 //Dump paragon.exe.
-                String fileToDump = Path.Combine(dstDiretory, "paragon.exe.dmp");
-                Program.Dump.MiniDumpToFile(fileToDump, "Paragon");
-
+                Process process = Process.GetProcessById(appInfo.BrowserInfo.Pid);// Process.GetProcessesByName("Paragon").First();
+                String fileToDump = Path.Combine(dstDiretory, process.ProcessName+".dmp");
+                MemoryDump.MiniDumpToFile(fileToDump, process);
+                
                 //Dump Paragon.Renderer.
-                fileToDump = Path.Combine(dstDiretory, "Paragon.Renderer.exe.dmp");
-                Program.Dump.MiniDumpToFile(fileToDump, "Paragon.Renderer");
-
+                process = Process.GetProcessById(appInfo.RenderInfo.Pid);
+                fileToDump = Path.Combine(dstDiretory, process.ProcessName + ".dmp");
+                MemoryDump.MiniDumpToFile(fileToDump, process);
+               
                 String packagePath = Path.Combine(dstDiretory, "package");
                 
                 //delete temp directory. Ensure cleanup.
